v2/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "portfolio-tracker"
version = "2.0.0"
description = "Portfolio Tracker v2 - Analytics tool for MultiWalk/MultiCharts systematic traders"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.35.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "numba>=0.59.0",
    "scipy>=1.13.0",
    "plotly>=5.22.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "pyxlsb>=1.0.10",
    "requests>=2.32.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "hypothesis>=6.100.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.hatch.build.targets.wheel]
packages = ["core", "ui"]
